Use @babel/transform-class-properties.
