+++ buildprocess/configureWebpack.js
@@ -3,6 +3,7 @@ const CopyPlugin = require("copy-webpack-plugin");
 const ForkTsCheckerWebpackPlugin = require("fork-ts-checker-webpack-plugin");
 const ForkTsCheckerNotifierWebpackPlugin = require("fork-ts-checker-notifier-webpack-plugin");
 const webpack = require("webpack");
+const defaultBabelLoader = require("./defaultBabelLoader");
 
 /**
  * Supplements the given webpack config with options required to build TerriaJS
@@ -123,31 +124,6 @@ function configureWebpack({
     }
   });
 
-  config.devServer = config.devServer || {
-    stats: "minimal",
-    port: 3003,
-    open: true,
-    contentBase: "wwwroot/",
-    proxy: {
-      "*": {
-        target: "http://localhost:3001",
-        bypass: function (req, res, proxyOptions) {
-          if (
-            req.url.indexOf("/proxy") < 0 &&
-            req.url.indexOf("/proj4lookup") < 0 &&
-            req.url.indexOf("/convert") < 0 &&
-            req.url.indexOf("/proxyabledomains") < 0 &&
-            req.url.indexOf("/errorpage") < 0 &&
-            req.url.indexOf("/init") < 0 &&
-            req.url.indexOf("/serverconfig") < 0
-          ) {
-            return req.originalUrl;
-          }
-        }
-      }
-    }
-  };
-
   config.plugins = config.plugins || [];
 
   // Do not import momentjs locale files
@@ -267,31 +243,5 @@ function configureWebpack({
 
   return config;
 }
-
-const defaultBabelLoader = ({ devMode }) => ({
-  loader: "babel-loader",
-  options: {
-    cacheDirectory: true,
-    sourceMaps: !!devMode,
-    presets: [
-      [
-        "@babel/preset-env",
-        {
-          corejs: 3,
-          useBuiltIns: "usage"
-        }
-      ],
-      ["@babel/preset-react", { runtime: "automatic" }],
-      ["@babel/preset-typescript", { allowNamespaces: true }]
-    ],
-    plugins: [
-      ["@babel/plugin-proposal-decorators", { legacy: true }],
-      "babel-plugin-styled-components"
-    ],
-    assumptions: {
-      setPublicClassFields: false
-    }
-  }
-});
 
 module.exports = configureWebpack;
